requirements.txt
colorlog==6.8.2
discord.py==2.2.2
python-dotenv==1.0.1
